Render project READMEs through the blog's component pipeline
READMEs now render via rehype-react with the blog's shared component map
instead of a sanitized HTML string, so code blocks use the CodeBlock
component (title bar, copy button, line numbers) and mermaid fences render
as live MermaidDiagram diagrams, matching blog prose. Extracts the shared
map to prose-components.tsx and drops the bespoke .readme-code CSS.
