+++ package.json
@@ -35,6 +35,7 @@
     "rehype-katex": "^7.0.1",
     "rehype-katex-notranslate": "^1.1.4",
     "rehype-raw": "^7.0.0",
+    "rehype-react": "^8.0.0",
     "rehype-sanitize": "^6.0.0",
     "rehype-slug": "^6.0.0",
     "rehype-stringify": "^10.0.1",
